componetização do userPage
