+++ src/pages/UserPage/User/style.module.css
@@ -1,0 +1,57 @@
+home {
+  background-color: #00c5f6;
+}
+.home-container {
+  width: 100%;
+  height: 100vh;
+  display: flex;
+  align-items: ;
+  flex-direction: column;
+  background-color: black;
+}
+.balance {
+  background-color: #1c1d21;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-evenly;
+  align-items: center;
+  width: 70vw;
+  height: 40vh;
+  border-radius: 1vw;
+  border: none;
+}
+.balance-container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  width: 90%;
+  height: 22vh;
+  background-color: #1c1d21;
+  border-radius: 5px;
+  border-radius: 1vw;
+}
+.balance-container-total {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: space-evenly;
+  background-color: #131519;
+}
+.balance-container-img {
+  display: flex;
+}
+.balance-container-value {
+  padding-left: 1vw;
+}
+.balance-container-available {
+  width: 30vw;
+  height: 18vh;
+  margin-top: 1vw;
+  border-radius: 1vw;
+  border: none;
+  background-color: #1c1d21;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
