--- src/pages/UserPage/style.module.css
@@ -1,14 +1,0 @@
-home {
-  background-color: #00c5f6;
-}
-.home-container {
-  width: 100%;
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-  background-color: black;
-}
-
-.balance-container {
-  border: solid blue;
-}
